Clean up readNetConfig, implement ClientConfig.
Split the readNetConfig into a version that accepts strings and
another version that accepts QXmlStreamReader. The latter is used
to recurse when parsing xml files (like for example, when parsing
for a config that contains net config elements, you can just call
readNetConfig with the XML stream pointing to the config.

Also, I fixed the way the parsing is done.... the previous one would
fail if the order of the nested tags were switched around...

The initial implementation of ClientConfig is just the same as netConfig.