Fix C++ Hash detection.
CMake's try_compile has troubles using variables in for loops.
Somehow, they get cached so if I used a variable again, the
cached result is returned even if the source file is different...

So the solution is to generate a new variable per loop, this can
be done using substitution.