proj: Add contestTimer to ProjectorWindow.
Instead of making the controllers have the contest timer, the contest
timer now belongs to the projector window. The reason is that the
contest timer emits signals and if we want the controllers to handle
the signals.... well... we're gonna have to make all the controllers
inherit from QObject which may not be pretty.