Did some testing on the Checker class. It compiles. I have also done a few tests on it already, albeit not as organized as I wanted to, but just enough to make sure that as long as the Answer and Question structs are properly initialized with values, it would work as expected. I'd like to test it some more when I get the chance before Saturday.
Modified some parts of the Question struct so that it gives
the right results for the checkAnswer method for CHOOSE_ANY.

Proper comments describing each method and explaining usage
has not yet been written down. Nonetheless, it would be added
within the next few days. The important thing right now is that
the Checker class works properly.

- Nikolai Banasihan and Vernon Gutierrez